requirements.txt
ipykernel==6.29.4
langchain-core==0.2.0
langchain==0.2.0
langchain_community==0.2.0
langchain_openai==0.1.7
openai==1.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pypdf==4.2.0
faiss-cpu==1.8.0
click==8.1.7
pyplz~=0.11.1